Moved functions from app.rs to display.rs
create_main_window, create_render_pass, and create_vertex_buffer for now.
